WIP - use aws cred file only
